requirements.txt
Django==2.1
celery[redis]==4.2.1
channels==2.1.7
channels-redis==2.3.3
django-environ==0.4.5
djangorestframework==3.9.2
pyppeteer==0.0.25
requests==2.21.0
websockets==6.0